requirements.txt
requests==2.31.0
schedule==1.2.0
APScheduler==3.10.4
websocket-client==1.6.4
python-dotenv==1.0.0
feedparser==6.0.10
beautifulsoup4==4.12.2
Pillow==10.0.1
openai==1.3.5
newspaper3k==0.2.8
